MAINT: Rename subpackage _py to _ivp in integrate
