+++ scipy/integrate/__init__.py
@@ -77,8 +77,8 @@ from .odepack import *
 from .quadpack import *
 from ._ode import *
 from ._bvp import solve_bvp
-from ._py import (solve_ivp, OdeSolution, DenseOutput,
-                  OdeSolver, RK23, RK45, Radau, BDF, LSODA)
+from ._ivp import (solve_ivp, OdeSolution, DenseOutput,
+                   OdeSolver, RK23, RK45, Radau, BDF, LSODA)
 
 __all__ = [s for s in dir() if not s.startswith('_')]
 from numpy.testing import Tester
