+++ scipy/integrate/_ivp/base.py
