+++ scipy/integrate/_ivp/common.py
