+++ scipy/integrate/_ivp/lsoda.py
