+++ scipy/integrate/_ivp/radau.py
